Moved push of data out of ShuffleWriter and addressed other minor comments
